Refactors ResourceItem to it's own component and adding it to the list, favorites, and category resources components
