+++ src/features/FavoriteResources.tsx
@@ -2,24 +2,20 @@ import React, { useEffect, useState, useContext } from 'react';
 import Skeleton from 'react-loading-skeleton';
 import { Types } from '@osu-wams/lib';
 import { Card, CardHeader, CardContent, CardFooter, CardIcon } from '../ui/Card';
-import { List, ListItem, ListItemContentLinkSVG, ListItemContentLinkName } from '../ui/List';
+import { List } from '../ui/List';
 import { faHeart } from '@fortawesome/pro-light-svg-icons';
-import { ThemeContext } from 'src/theme';
 import { useResources } from '@osu-wams/hooks';
 import { UserContext } from 'src/App';
-import { TrendingEvent } from './resources/GATrendingResource';
 import { Event } from 'src/util/gaTracking';
-import { IconLookup } from './resources/resources-utils';
 import FailedState from 'src/ui/FailedState';
 import { InternalLink } from 'src/ui/Link';
 import { EmptyState } from 'src/ui/EmptyStates';
-import { ResourceItem } from 'src/features/ResourcesCard';
+import { ResourceItem } from './resources/ResourceItem';
 
 export const FavoriteResources = () => {
   const user = useContext(UserContext);
   const res = useResources();
   const [favoriteResources, setFavoriteResources] = useState<Types.Resource[]>([]);
-  const themeContext = useContext(ThemeContext);
 
   useEffect(() => {
     if (user.data.favoriteResources && res.data.length > 0) {
@@ -47,19 +43,6 @@ export const FavoriteResources = () => {
                 resource={resource}
                 event={() => Event('favorite-resources-card', resource.title)}
               />
-              // <ListItem key={resource.id}>
-              //   <ListItemContentLinkSVG
-              //     href={resource.link}
-              //     target="_blank"
-              //     onClick={() => {
-              //       Event('favorite-resources-card', resource.title);
-              //       TrendingEvent(resource, user.data);
-              //     }}
-              //   >
-              //     {IconLookup(resource.iconName, themeContext.features.resources.icon.color)}
-              //     <ListItemContentLinkName>{resource.title}</ListItemContentLinkName>
-              //   </ListItemContentLinkSVG>
-              // </ListItem>
             ))}
           </List>
         )}
